Get the gridify layouts

Lists out all the layout functions exported from the gridify package.

Usage

get_layouts()

Value

A vector listing out the names of the layout functions from the gridify package.

See also

Examples

get_layouts()
#> [1] "complex_layout"       "simple_layout"        "pharma_layout_base"  
#> [4] "pharma_layout_A4"     "pharma_layout_letter"